resolution bug #2

+++ WebService/library-business/src/main/java/org/WebService/business/contract/manager/BookManager.java
@@ -7,13 +7,9 @@ import java.util.List;
 
 public interface BookManager {
 
-    Book getBook(long id);
-
+    Book getBook(Integer id);
     List<Book> getAllBooks();
     List<Book> getBookByName(String name);
-
     void createBook(Book book);
-
-
     List<Book> getBooksByAuthor(Author author);
 }
